01.02.20 update

+++ database/migrations/2020_01_02_200601_add-user-buffer.php
@@ -19,10 +19,10 @@ class AddUserBuffer extends Migration
             $table->string('surname');
             $table->string('patronymic');
             $table->string('email');
-            $table->integer('faculty');
-            $table->integer('departament');
-            $table->integer('degree');
-            $table->integer('post');
+            $table->unsignedBigInteger('faculty');
+            $table->unsignedBigInteger('departament');
+            $table->unsignedBigInteger('degree');
+            $table->unsignedBigInteger('post');
             $table->string('password_no_hash');
             $table->timestampsTz();
         });
